[caliper] orchestration: postprocess: move logging code to postprocess_logging
